Added VC 2010 project for ICQ icons (patch by FREAK_THEMIGHTY) Fixed Jabber compiler warning
git-svn-id: http://miranda.googlecode.com/svn/branches/stable@13682 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/protocols/JabberG/jabber_secur.cpp
@@ -442,7 +442,7 @@ char* TPlainAuth::getInitialRequest()
 	mir_free( uname );
 	mir_free( passw );
 
-	return JabberBase64Encode( toEncode, size );
+	return JabberBase64Encode( toEncode, (int)size );
 }
 
 /////////////////////////////////////////////////////////////////////////////////////////
